+++ test-utils/src/main/java/com.bitmovin.analytics.test.utils/PlayerSettings.kt
@@ -1,3 +1,3 @@
 package com.bitmovin.analytics.test.utils
 
-data class PlayerSettings(val isMuted: Boolean, val isAutoPlayEnabled: Boolean? = null)
+data class PlayerSettings(val isMuted: Boolean, val isAutoPlayEnabled: Boolean = false)